Migrate GstGLVideoItem to Qt6 with fixes
In particular
* uses new Qt6 qmlsink gstreamer plugin code (switch to new submodule repo)
* removes unnecessary dependency of QGC to VideoReceiverApp
* removes dependency of VideoReceiver to Settings submodule (uses only enum from Settings)
* removes registering of GstGLVideoItem in GStreamer.cc (it's registered in own code)
* minor cleanup and fixes in VideoReceiverApp
* fixes dragging map when resizing/clicking on Pip Window
