+++ .gitmodules
@@ -9,7 +9,7 @@
 	url = https://github.com/KDAB/android_openssl
 [submodule "libs/qmlglsink/gst-plugins-good"]
 	path = libs/qmlglsink/gst-plugins-good
-	url = https://github.com/mavlink/gst-plugins-good.git
+    url = https://github.com/zdanek/qgc-gst-plugins-good.git
 [submodule "libs/xz-embedded"]
 	path = libs/xz-embedded
 	url = https://github.com/Auterion/xz-embedded.git
