+++ libs/CMakeLists.txt
@@ -10,5 +10,7 @@ if (GST_FOUND)
     add_subdirectory(qmlglsink)
 endif()
 
-#set(BUILD_SHARED_LIBS OFF CACHE BOOL "Force qmdnsengine to build as static" FORCE)
-#add_subdirectory(qmdnsengine)
+if (NOT QT6_DISABLE_DNSENGINE)
+    set(BUILD_SHARED_LIBS OFF CACHE BOOL "Force qmdnsengine to build as static" FORCE)
+    add_subdirectory(qmdnsengine)
+endif ()
