+++ src/CMakeLists.txt
@@ -10,7 +10,6 @@ include_directories(
         ${CMAKE_CURRENT_SOURCE_DIR}
         ${CMAKE_CURRENT_BINARY_DIR}
         ${Qt6Location_PRIVATE_INCLUDE_DIRS}
-        VideoManager
 )
 
 #######################################################
