+++ src/QmlControls/DeadMouseArea.qml
@@ -4,7 +4,7 @@ import QtQuick.Controls
 MouseArea {
     preventStealing:true
     hoverEnabled:   true
-    onWheel:    (wheel) => (wheel) => { wheel.accepted = true; }
+    onWheel:    (wheel) => { wheel.accepted = true; }
     onPressed:  (mouse) => { mouse.accepted = true; }
     onReleased: (mouse) => { mouse.accepted = true; }
 }
