+++ src/Settings/VideoDecoderOptions.h
@@ -1,0 +1,22 @@
+/****************************************************************************
+*
+ * (c) 2009-2023 QGROUNDCONTROL PROJECT <http://www.qgroundcontrol.org>
+ *
+ * QGroundControl is licensed according to the terms in the file
+ * COPYING.md in the root of the source code directory.
+ *
+ ****************************************************************************/
+
+#ifndef VIDEODECODEROPTIONS_H
+#define VIDEODECODEROPTIONS_H
+
+enum VideoDecoderOptions {
+    ForceVideoDecoderDefault = 0,
+    ForceVideoDecoderSoftware,
+    ForceVideoDecoderNVIDIA,
+    ForceVideoDecoderVAAPI,
+    ForceVideoDecoderDirectX3D,
+    ForceVideoDecoderVideoToolbox,
+};
+
+#endif //VIDEODECODEROPTIONS_H
